Moving model files into lib/model/
