+++ src/bin/generate-test.ts
@@ -6,8 +6,8 @@ import { DataSet } from '../lib/data/set';
 import { Engine } from "../lib/engine/engine";
 import { Hubspot } from '../lib/hubspot';
 import { Logger } from '../lib/log';
-import { License } from '../lib/marketplace/model/license';
-import { Transaction } from '../lib/marketplace/model/transaction';
+import { License } from '../lib/model/license';
+import { Transaction } from '../lib/model/transaction';
 import { abbrActionDetails, abbrEventDetails } from '../tests/deal-generator/utils';
 
 function TEMPLATE({ runDealGenerator, GROUP, RECORDS, EVENTS, ACTIONS }: any) {
