+++ src/lib/engine/engine.ts
@@ -1,17 +1,17 @@
 import chalk from "chalk";
 import { Data } from "../data/set";
 import { Hubspot } from "../hubspot";
-import { CompanyManager } from "../hubspot/model/company";
-import { ContactManager } from "../hubspot/model/contact";
-import { DealManager } from "../hubspot/model/deal";
 import { Logger } from "../log";
 import { Table } from "../log/table";
 import { Tallier } from "../log/tallier";
-import { License } from "../marketplace/model/license";
-import { getEmailsForRecord } from "../marketplace/model/record";
-import { Transaction } from "../marketplace/model/transaction";
 import { buildAndVerifyStructures } from "../marketplace/structure";
 import * as validation from "../marketplace/validation";
+import { CompanyManager } from "../model/company";
+import { ContactManager } from "../model/contact";
+import { DealManager } from "../model/deal";
+import { License } from "../model/license";
+import { getEmailsForRecord } from "../model/record";
+import { Transaction } from "../model/transaction";
 import { formatMoney, formatNumber } from "../util/formatters";
 import { deriveMultiProviderDomainsSet } from "./all-free-email-providers";
 import { identifyAndFlagContactTypes } from "./contacts/contact-types";
