+++ src/lib/engine/summary.ts
@@ -1,5 +1,5 @@
-import { Deal } from "../hubspot/model/deal";
 import { Table } from "../log/table";
+import { Deal } from "../model/deal";
 import { formatMoney, formatNumber } from "../util/formatters";
 import { isPresent } from "../util/helpers";
 import { Engine } from "./engine";
