+++ src/lib/marketplace/structure.ts
@@ -1,9 +1,9 @@
 import { Logger } from "../log";
 import { Table } from "../log/table";
+import { License } from "../model/license";
+import { Transaction, TransactionData } from "../model/transaction";
 import { formatMoney } from "../util/formatters";
 import { isPresent } from "../util/helpers";
-import { License } from "./model/license";
-import { Transaction, TransactionData } from "./model/transaction";
 
 export function buildAndVerifyStructures(licenses: License[], transactions: Transaction[], log?: Logger) {
   return new Structurer(log).buildAndVerify(licenses, transactions);
