+++ src/lib/marketplace/validation.ts
@@ -1,9 +1,9 @@
 import assert from "assert";
 import util from "util";
 import { Logger } from '../log';
+import { License, LicenseData } from "../model/license";
+import { Transaction } from "../model/transaction";
 import { AttachableError } from '../util/errors';
-import { License, LicenseData } from "./model/license";
-import { Transaction } from "./model/transaction";
 
 export function removeApiBorderDuplicates(licenses: readonly License[]) {
   const groups: { [addonLicenseId: string]: License[] } = {};
